requirements.txt
torchvision
hydra-core
tqdm